+++ ContentPatcher/docs/release-notes.md
@@ -9,6 +9,9 @@ When releasing a format change, don't forget to update the smapi.io/json schema!
 
 
 -->
+## Upcoming release
+* Reimplemented internal `Patch.LastChangedTick` property for backwards compatibility with Content Patcher Animations.
+
 ## 2.5.2
 Released 07 February 2025 for SMAPI 4.1.10 or later.